reporting stack support one user

+++ src/control-plane/backend/lambda/api/service/reporting.ts
@@ -990,7 +990,11 @@ async function _cleanDatasets(quickSight: QuickSight) {
   }
 
   for (const dataSetSummary of dataSetSummaries) {
-    if (dataSetSummary.DataSetId?.startsWith(QUICKSIGHT_TEMP_RESOURCE_NAME_PREFIX)) {
+    if (
+      dataSetSummary.DataSetId?.startsWith(QUICKSIGHT_TEMP_RESOURCE_NAME_PREFIX)
+      && dataSetSummary.CreatedTime !== undefined
+      && (new Date().getTime() - dataSetSummary.CreatedTime.getTime()) > 60 * 60 * 1000
+    ) {
       const dataSetId = dataSetSummary.DataSetId;
       logger.info(`deleting data set: ${ dataSetId }`);
       const deletedRes = await quickSight.deleteDataSet({
@@ -1019,7 +1023,11 @@ async function _cleanAnalyses(quickSight: QuickSight) {
 
   for (const analysisSummary of analysisSummaries) {
     const analysisId = analysisSummary.AnalysisId;
-    if (analysisSummary.Status !== ResourceStatus.DELETED) {
+    if (analysisSummary.Status !== ResourceStatus.DELETED
+      && analysisSummary.CreatedTime !== undefined
+      && analysisSummary.AnalysisId?.startsWith(QUICKSIGHT_TEMP_RESOURCE_NAME_PREFIX) &&
+      (new Date().getTime() - analysisSummary.CreatedTime.getTime()) > 60 * 60 * 1000
+    ) {
       await quickSight.deleteAnalysis({
         AwsAccountId: awsAccountId,
         AnalysisId: analysisId,
@@ -1047,7 +1055,10 @@ async function _cleanedDashboard(quickSight: QuickSight) {
 
   for (const dashboardSummary of dashboardSummaries) {
     const dashboardId = dashboardSummary.DashboardId;
-    if (dashboardSummary.DashboardId?.startsWith(QUICKSIGHT_TEMP_RESOURCE_NAME_PREFIX)) {
+    if (dashboardSummary.DashboardId?.startsWith(QUICKSIGHT_TEMP_RESOURCE_NAME_PREFIX)
+      && dashboardSummary.CreatedTime !== undefined
+      && (new Date().getTime() - dashboardSummary.CreatedTime.getTime()) > 60 * 60 * 1000
+    ) {
       await quickSight.deleteDashboard({
         AwsAccountId: awsAccountId,
         DashboardId: dashboardId,
